[Doc] Enable generation of subdirectories to store the output
The files kinetics.h and onedim.h are present as high-level application 
include files and as lower-level implementation include files. Doxygen 
generates HTML documentation for all files in the same folder, and on 
case-insensitive file systems, the documetation for the application and 
the implementation files will shadow each other. This commit enables the 
option to put documentation files into subfolders in the root directory, 
hopefully separating these "duplicate" files