[Kinetics] Allow negative pre-exponential factor for falloff reactions
If both the high and low rate coefficients for a falloff or chemically
activated reaction are negative, then there is no problem evaluating the rate,
so it can be allowed in the same cases where a negative rate would be
allowed for an elementary reaction, i.e. if the presence of the negative
pre-exponential factors has been explicitly declared.